feat: use search parameter as input filter
Providing a search parameter when starting bwtui will now pre-fill the
filter field at the bottom with the given parameter value. This
provides the benefit of providing the same UX as if the user used '/'
and then typed the search term there so that they can also clear the
search and see all items again which wasn't possible before.